refactor: 프로바이더 아이콘 위젯으로 추출
